requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.5
djangorestframework==3.13.1
gunicorn==20.1.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
static3==0.7.0
whitenoise==6.2.0
